move sdram to its own module
